Move containers into containers directory.
